pyproject.toml
[project]
name = "ragproductionapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "inngest>=0.5.11",
    "llama-index-core>=0.14.7",
    "llama-index-readers-file>=0.5.4",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
